// types.ts
export enum View {
  LANDING = 'LANDING',
  DASHBOARD = 'DASHBOARD',
  ADD_CONNECTION = 'ADD_CONNECTION',
  PROFILE = 'PROFILE',
  CIRCLE = 'CIRCLE',
  JOURNAL = 'JOURNAL',
  CALENDAR = 'CALENDAR',
  SETTINGS = 'SETTINGS',
  PRIVACY = 'PRIVACY',
  TERMS = 'TERMS',
  GUIDE = 'GUIDE',
  ABOUT = 'ABOUT'
}